Rename wrapper files.
